requirements.txt
# Core agent framework
langgraph>=0.2
langchain-core
langchain-google-genai>=3.0
typing_extensions
amadeus

# AI/LLM
pydantic-ai>=1.9
pydantic-ai-slim>=1.9
google-genai>=1.7.0

# Data validation & API
pydantic>=2.0
fastapi>=0.104
uvicorn[standard]>=0.24  # ASGI server for FastAPI

# HTTP client
httpx>=0.27
tenacity>=8.2

# UI (if needed)
streamlit>=1.36
watchfiles>=0.21

# Environment variables
python-dotenv>=1.0.0

# Session storage
redis>=5.0.0
hiredis>=2.2.0

# AWS integration (optional, for Secrets Manager)
boto3>=1.34.0

# Testing & Linting
pytest>=7.4
pytest-asyncio>=0.21
pytest-cov>=4.1
ruff>=0.14